Reorganize ANTLR generated code as per recommendations in ANTLR README
