+++ antlr/generator.go
@@ -1,0 +1,3 @@
+package antlr
+
+//go:generate ./generate.sh
