+++ flows/definition/legacy/expressions/migrate.go
@@ -7,12 +7,11 @@ import (
 	"strconv"
 	"strings"
 
+	"github.com/antlr/antlr4/runtime/Go/antlr/v4"
+	gen "github.com/nyaruka/goflow/antlr/gen/excellent1"
 	"github.com/nyaruka/goflow/envs"
 	"github.com/nyaruka/goflow/excellent"
 	"github.com/nyaruka/goflow/flows"
-	"github.com/nyaruka/goflow/flows/definition/legacy/gen"
-
-	"github.com/antlr/antlr4/runtime/Go/antlr/v4"
 )
 
 // ContextTopLevels are the allowed top-level identifiers in legacy expressions, i.e. @contact.bar is valid but @foo.bar isn't
